# order-service/src/main/resources/openapi/order-api.yaml
openapi: 3.0.3
info:
  title: Order Service API
  description: Order CRUD, validation, invoice, and status APIs
  version: 1.0.0

servers:
  - url: http://localhost:8081/orders
    description: Local

tags:
  - name: Orders
  - name: Health

paths:
  /v1/health:
    get:
      tags:
        - Health
      summary: Health probe
      operationId: health
      responses:
        "200":
          description: Service is up
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  type: string

  /v1/orders:
    get:
      tags:
        - Orders
      summary: List orders
      operationId: listOrders
      parameters:
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/OrderStatus"

        - name: page
          in: query
          schema:
            type: integer
            default: 0

        - name: size
          in: query
          schema:
            type: integer
            default: 20

      responses:
        "200":
          description: Paginated orders
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PageOrderResponse"

    post:
      tags:
        - Orders
      summary: Create order
      operationId: createOrder
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"

      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderRequest"

      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderResponse"

        "400":
          description: Missing Idempotency-Key or validation error

  /v1/orders/{orderId}:
    get:
      tags:
        - Orders
      summary: Get order by ID
      operationId: getOrder
      parameters:
        - $ref: "#/components/parameters/OrderId"

      responses:
        "200":
          description: Order found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderResponse"

        "404":
          description: Order not found

  /v1/orders/{orderId}/validate:
    post:
      tags:
        - Orders
      summary: Validate order
      operationId: validateOrder
      parameters:
        - $ref: "#/components/parameters/OrderId"
        - $ref: "#/components/parameters/IdempotencyKey"

      responses:
        "200":
          description: Validation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationResultResponse"

  /v1/orders/{orderId}/invoice:
    post:
      tags:
        - Orders
      summary: Generate invoice
      operationId: generateInvoice
      parameters:
        - $ref: "#/components/parameters/OrderId"
        - $ref: "#/components/parameters/IdempotencyKey"

      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InvoiceRequest"

      responses:
        "201":
          description: Invoice created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvoiceResponse"

  /v1/orders/{orderId}/status:
    patch:
      tags:
        - Orders
      summary: Update order status
      operationId: updateStatus
      parameters:
        - $ref: "#/components/parameters/OrderId"
        - $ref: "#/components/parameters/IdempotencyKey"

      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StatusUpdateRequest"

      responses:
        "200":
          description: Status updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderResponse"

        "409":
          description: Illegal status transition

components:
  parameters:
    OrderId:
      name: orderId
      in: path
      required: true
      schema:
        type: string
        format: uuid

    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: Client-supplied unique key for mutating requests (24h Redis dedup)
      schema:
        type: string
        format: uuid

  schemas:
    OrderStatus:
      type: string
      enum:
        - CREATED
        - VALIDATED
        - PROCESSING
        - PAYMENT_PENDING
        - FRAUD_REVIEW
        - PAID
        - SHIPPED
        - COMPLETED
        - COMPENSATING
        - CANCELLED
        - FAILED
        - PAYMENT_TIMEOUT
        - REJECTED

    CreateOrderRequest:
      type: object
      required:
        - customerId
        - items
        - shippingAddress
      properties:
        customerId:
          type: string

        items:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/OrderItemRequest"

        shippingAddress:
          $ref: "#/components/schemas/ShippingAddressRequest"

        currency:
          type: string
          example: INR

    OrderItemRequest:
      type: object
      required:
        - sku
        - quantity
        - unitPrice
      properties:
        sku:
          type: string

        quantity:
          type: integer
          minimum: 1

        unitPrice:
          type: number
          format: double
          minimum: 0.01

    ShippingAddressRequest:
      type: object
      required:
        - line1
        - city
        - state
        - postalCode
        - country
      properties:
        line1:
          type: string

        line2:
          type: string

        city:
          type: string

        state:
          type: string

        postalCode:
          type: string

        country:
          type: string

    OrderResponse:
      type: object
      properties:
        orderId:
          type: string

        customerId:
          type: string

        status:
          $ref: "#/components/schemas/OrderStatus"

        totalAmount:
          type: number

        currency:
          type: string

        workflowInstanceKey:
          type: string
          nullable: true

        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItemRequest"

        shippingAddress:
          $ref: "#/components/schemas/ShippingAddressRequest"

        createdAt:
          type: string
          format: date-time

        updatedAt:
          type: string
          format: date-time

    PageOrderResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/OrderResponse"

        page:
          type: integer

        size:
          type: integer

        totalElements:
          type: integer
          format: int64

        totalPages:
          type: integer

    ValidationResultResponse:
      type: object
      properties:
        valid:
          type: boolean

        validItems:
          type: boolean

        validAddress:
          type: boolean

        errors:
          type: array
          items:
            type: string

    InvoiceRequest:
      type: object
      required:
        - paymentId
      properties:
        orderId:
          type: string

        paymentId:
          type: string

    InvoiceResponse:
      type: object
      properties:
        invoiceId:
          type: string

        orderId:
          type: string

        pdfUrl:
          type: string

        amount:
          type: number

        issuedAt:
          type: string
          format: date-time

    StatusUpdateRequest:
      type: object
      required:
        - status
      properties:
        status:
          $ref: "#/components/schemas/OrderStatus"

        reason:
          type: string